+++ apps/web/package.json
@@ -19,6 +19,7 @@
     "@t3-oss/env-nextjs": "^0.9.2",
     "@tanstack/react-query": "^5.40.0",
     "@tanstack/react-query-devtools": "^5.40.0",
+    "argon2-browser": "^1.18.0",
     "class-variance-authority": "^0.7.0",
     "jotai": "^2.8.3",
     "lucide-react": "^0.344.0",
@@ -33,6 +34,7 @@
     "@buildit/eslint-config": "workspace:*",
     "@buildit/tailwind-config": "workspace:*",
     "@buildit/tsconfig": "workspace:*",
+    "@types/argon2-browser": "^1.18.4",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
