+++ apps/web/src/app/sign-in/page.tsx
@@ -1,7 +1,5 @@
-import Link from "next/link";
-
 import { OauthButton } from "@/components/forms/oauth-buttons";
-import SignUpForm from "@/components/forms/sign-up-form";
+import SignInForm from "@/components/forms/sign-in-form";
 
 export const runtime = "edge";
 
@@ -10,25 +8,19 @@ export default function RegisterPage() {
     <div className="mx-auto flex h-dvh items-center justify-center">
       <div className="bg-muted border-subtle w-fit rounded-lg border p-16">
         <h1 className="font-cal text-emphasis text-2xl">
-          Create your BuildIt
+          Signin to your BuildIt
           <br /> account
         </h1>
-        <p className="text-subtle">
+        <p className="text-subtle text-sm">
           Free for individuals. Team plan for <br /> collaborative features.
         </p>
-        <SignUpForm />
+        <SignInForm />
         <div className="my-4 flex w-full items-center">
           <div className="bg-emphasis h-[1px] w-full" />
           <p className="text-subtle px-2 text-sm">OR</p>
           <div className="bg-emphasis h-[1px] w-full" />
         </div>
         <OauthButton />
-        <p className="text-subtle mt-4 text-xs">
-          Already have an account?{" "}
-          <Link href={"/auth/signin"} className="text-default font-semibold">
-            Sign in
-          </Link>
-        </p>
       </div>
     </div>
   );
