+++ apps/web/src/components/forms/sign-in-form.tsx
@@ -1,9 +1,8 @@
 "use client";
 
+import { useRouter } from "next/navigation";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
-import { toast } from "sonner";
 import { z } from "zod";
 
 import {
@@ -17,52 +16,36 @@ import {
   Input,
 } from "@buildit/ui";
 
-import { login } from "@/lib/actions/auth/login";
-
 const SignInSchema = z.object({
+  username: z
+    .string()
+    .min(3, "Username must be at least 3 characters long")
+    .max(20, "Username cannot exceed 20 characters")
+    .regex(/^[a-zA-Z0-9_]*$/, {
+      message: "Only letters, numbers, and underscores are allowed.",
+    }),
   email: z.string().email("Invalid email format"),
   password: z.string().min(8, "Password must be at least 8 characters long"),
 });
 
 export default function SignInForm() {
+  const router = useRouter();
   const form = useForm<z.infer<typeof SignInSchema>>({
     resolver: zodResolver(SignInSchema),
     defaultValues: {
       email: "",
-      password: "",
-    },
-  });
-
-  const mutation = useMutation({
-    mutationKey: ["login"],
-    mutationFn: (values: z.infer<typeof SignInSchema>) =>
-      login({
-        email: values.email,
-        password: values.password,
-      }),
-
-    onSuccess: (res) => {
-      if (res.error) {
-        toast.error(res.error);
-        form.reset();
-      } else {
-        toast.success(res.success);
-      }
-    },
-    onError: (error) => {
-      toast.error(error.message);
     },
   });
 
   const onSubmit = (values: z.infer<typeof SignInSchema>) => {
-    mutation.mutate(values);
+    console.log(values);
   };
 
   return (
     <Form {...form}>
       <form
         onSubmit={form.handleSubmit(onSubmit)}
-        className="w-[296px] space-y-6"
+        className="mt-6 w-[296px] space-y-6"
       >
         <FormField
           control={form.control}
@@ -72,7 +55,7 @@ export default function SignInForm() {
               <FormLabel>Email</FormLabel>
               <FormControl>
                 <Input
-                  placeholder="Enter email"
+                  placeholder="onboarding@example.com"
                   {...field}
                   className="bg-white"
                 />
@@ -81,26 +64,9 @@ export default function SignInForm() {
             </FormItem>
           )}
         />
-        <FormField
-          control={form.control}
-          name="password"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Password</FormLabel>
-              <FormControl>
-                <Input
-                  type="password"
-                  placeholder="Enter password"
-                  {...field}
-                  className="bg-white"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
+
         <Button type="submit" className="w-full">
-          Login
+          Sign In
         </Button>
       </form>
     </Form>
