--- apps/web/src/components/forms/sign-up-form.tsx
@@ -1,136 +1,0 @@
-"use client";
-
-import { useRouter } from "next/navigation";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation } from "@tanstack/react-query";
-import { useForm } from "react-hook-form";
-import { toast } from "sonner";
-import { z } from "zod";
-
-import {
-  Button,
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-  Input,
-} from "@buildit/ui";
-
-import { createAccount } from "@/lib/actions/auth/create-account";
-
-const SignUpSchema = z.object({
-  username: z
-    .string()
-    .min(3, "Username must be at least 3 characters long")
-    .max(20, "Username cannot exceed 20 characters")
-    .regex(/^[a-zA-Z0-9_]*$/, {
-      message: "Only letters, numbers, and underscores are allowed.",
-    }),
-  email: z.string().email("Invalid email format"),
-  password: z.string().min(8, "Password must be at least 8 characters long"),
-});
-
-export default function SignUpForm() {
-  const router = useRouter();
-  const form = useForm<z.infer<typeof SignUpSchema>>({
-    resolver: zodResolver(SignUpSchema),
-    defaultValues: {
-      username: "",
-      email: "",
-      password: "",
-    },
-  });
-
-  const mutation = useMutation({
-    mutationKey: ["createAccount"],
-    mutationFn: (values: z.infer<typeof SignUpSchema>) =>
-      createAccount({
-        username: values.username,
-        email: values.email,
-        password: values.password,
-      }),
-    onSuccess: (res) => {
-      if (res.error) {
-        toast.error(res.error);
-        form.reset();
-      } else {
-        toast.success(res.success);
-        router.push("/auth/signin");
-      }
-    },
-    onError: (error) => {
-      toast.error(error.message);
-    },
-  });
-
-  const onSubmit = (values: z.infer<typeof SignUpSchema>) => {
-    mutation.mutate(values);
-  };
-
-  return (
-    <Form {...form}>
-      <form
-        onSubmit={form.handleSubmit(onSubmit)}
-        className="mt-6 w-[296px] space-y-6"
-      >
-        <FormField
-          control={form.control}
-          name="username"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Username</FormLabel>
-              <FormControl>
-                <Input
-                  placeholder="Enter username"
-                  {...field}
-                  className="bg-white"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="email"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Email</FormLabel>
-              <FormControl>
-                <Input
-                  placeholder="Enter email"
-                  {...field}
-                  className="bg-white"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="password"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Password</FormLabel>
-              <FormControl>
-                <Input
-                  type="password"
-                  placeholder="Enter password"
-                  {...field}
-                  className="bg-white"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <Button type="submit" className="w-full">
-          Create Account
-        </Button>
-      </form>
-    </Form>
-  );
-}
